move functionalities from `support.py` module to `bank_lines.utils.py`
