+++ src/dfastbe/support.py
@@ -234,7 +234,7 @@ def get_slices_core(
         - If `bmax1` is True, intersections beyond the endpoint of the segment
           are ignored.
     """
-    a, b, slices = get_slices_ab(
+    a, b, slices = _get_slices_ab(
         mesh_data.x_edge_coords[edges, 0],
         mesh_data.y_edge_coords[edges, 0],
         mesh_data.x_edge_coords[edges, 1],
@@ -250,7 +250,7 @@ def get_slices_core(
     return a, b, edges
 
 
-def get_slices_ab(
+def _get_slices_ab(
     X0: numpy.ndarray,
     Y0: numpy.ndarray,
     X1: numpy.ndarray,
@@ -341,16 +341,16 @@ def move_line(
         Nx2 array containing the x- and y-coordinates of the moved line.
     """
     if right_bank:
-        xylines_new = move_line_right(xylines, erosion_distance)
+        xylines_new = _move_line_right(xylines, erosion_distance)
     else:
         xylines_rev = xylines[::-1, :]
         dn_rev = erosion_distance[::-1]
-        xylines_new_rev = move_line_right(xylines_rev, dn_rev)
+        xylines_new_rev = _move_line_right(xylines_rev, dn_rev)
         xylines_new = xylines_new_rev[::-1, :]
     return xylines_new
 
 
-def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> numpy.ndarray:
+def _move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> numpy.ndarray:
     """
     Shift a line of a variable distance sideways (positive shift to the right).
 
@@ -380,8 +380,8 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
 
     xylines_new = numpy.zeros((100, 2))
     xylines_new[0] = xylines[0] + nxy[0]
-    ixy, xylines_new = add_point(0, xylines_new, xylines[1] + nxy[0])
-    ixy, xylines_new = add_point(ixy, xylines_new, xylines[1])
+    ixy, xylines_new = _add_point(0, xylines_new, xylines[1] + nxy[0])
+    ixy, xylines_new = _add_point(ixy, xylines_new, xylines[1])
 
     verbose = False
     prec = 0.000001
@@ -496,7 +496,7 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
                 pass
             else:
                 # check for intersection
-                a2, b2, slices2 = get_slices_ab(
+                a2, b2, slices2 = _get_slices_ab(
                     X0,
                     Y0,
                     X1,
@@ -553,7 +553,7 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
             for n2 in range(min(nedges, 2), -1, -1):
                 if verbose:
                     print("  adding point {}".format(poly[n2]))
-                ixy1, xylines_new = add_point(ixy1, xylines_new, poly[n2])
+                ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
             ixy = ixy1
 
         else:
@@ -594,7 +594,7 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
                         for n2 in range(n_last, n[i], -1):
                             if verbose:
                                 print("  adding new point {}".format(poly[n2]))
-                            ixy1, xylines_new = add_point(ixy1, xylines_new, poly[n2])
+                            ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
                     else:
                         if verbose:
                             print("  existing line is outside the new polygon")
@@ -605,13 +605,13 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
                                         xytmp[s2 - ixytmp + 1]
                                     )
                                 )
-                            ixy1, xylines_new = add_point(
+                            ixy1, xylines_new = _add_point(
                                 ixy1, xylines_new, xytmp[s2 - ixytmp + 1]
                             )
                     pnt_intersect = poly[n[i]] + b[i] * (poly[n[i] + 1] - poly[n[i]])
                     if verbose:
                         print("  adding intersection point {}".format(pnt_intersect))
-                    ixy1, xylines_new = add_point(ixy1, xylines_new, pnt_intersect,)
+                    ixy1, xylines_new = _add_point(ixy1, xylines_new, pnt_intersect, )
                     n_last = n[i]
                     s_last = s[i]
                     if a[i] < prec:
@@ -648,14 +648,14 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
                 for n2 in range(n_last, -1, -1):
                     if verbose:
                         print("  adding new point {}".format(poly[n2]))
-                    ixy1, xylines_new = add_point(ixy1, xylines_new, poly[n2])
+                    ixy1, xylines_new = _add_point(ixy1, xylines_new, poly[n2])
             else:
                 if verbose:
                     print("  existing line is inside the new polygon")
                 for s2 in range(s_last, len(xytmp) + ixytmp - 1):
                     if verbose:
                         print("  re-adding old point {}".format(xytmp[s2 - ixytmp + 1]))
-                    ixy1, xylines_new = add_point(
+                    ixy1, xylines_new = _add_point(
                         ixy1, xylines_new, xytmp[s2 - ixytmp + 1]
                     )
             ixy = ixy1
@@ -666,7 +666,7 @@ def move_line_right(xylines: numpy.ndarray, erosion_distance: numpy.ndarray) -> 
     return xylines_new
 
 
-def add_point(
+def _add_point(
     ixy1: int, xy_in: numpy.ndarray, point: numpy.ndarray
 ) -> Tuple[int, numpy.ndarray]:
     """
@@ -698,262 +698,6 @@ def add_point(
     else:
         xy_out = xy_in
     return ixy1, xy_out
-
-
-def sort_connect_bank_lines(
-    banklines: MultiLineString,
-    xykm: LineString,
-    right_bank: bool,
-) -> LineString:
-    """
-    Connect the bank line segments to bank lines.
-
-    Arguments
-    ---------
-    banklines : MultiLineString
-        Unordered set of bank line segments.
-    xykm : LineString
-        Array containing x,y,chainage values.
-    right_bank : bool
-        Flag indicating whether line is on the right (or not).
-
-    Returns
-    -------
-    bank : LineString
-        The detected bank line.
-    """
-
-    # convert MultiLineString into list of LineStrings that can be modified later
-    banklines_list = [line for line in banklines.geoms]
-
-    # loop over banklines and determine minimum/maximum projected length
-    # print("numpy init")
-    minlocs = numpy.zeros(len(banklines_list))
-    maxlocs = numpy.zeros(len(banklines_list))
-    lengths = numpy.zeros(len(banklines_list))
-    keep = lengths == 1
-    # print("loop {} bank lines".format(len(banklines_list)))
-    for i, bl in enumerate(banklines_list):
-        minloc = 1e20
-        maxloc = -1
-        for j, p in enumerate(bl.coords):
-            loc = xykm.project(Point(p))
-            if loc < minloc:
-                minloc = loc
-                minj = j
-            if loc > maxloc:
-                maxloc = loc
-                maxj = j
-        minlocs[i] = minloc  # at minj
-        maxlocs[i] = maxloc  # at maxj
-        if minj == maxj:
-            pass  # if minj == maxj then minloc == maxloc and thus lengths == 0 and will be removed anyway
-        elif bl.coords[0] == bl.coords[-1]:
-            # print(i,"cyclic", minloc, maxloc)
-            crd_numpy = numpy.array(bl.coords)
-            ncrd = len(crd_numpy)
-            if minj < maxj:
-                op1 = numpy.array(bl.coords[minj : maxj + 1])
-                op2 = numpy.zeros((ncrd + minj - maxj, 2))
-                op2[0 : ncrd - maxj - 1] = crd_numpy[maxj:-1]
-                op2[ncrd - maxj - 1 : ncrd - maxj + minj] = crd_numpy[: minj + 1]
-                op2 = op2[::-1]
-            else:  # minj > maxj
-                op1 = numpy.array(bl.coords[maxj : minj + 1][::-1])
-                op2 = numpy.zeros((ncrd + maxj - minj, 2))
-                op2[0 : ncrd - minj - 1] = crd_numpy[minj:-1]
-                op2[ncrd - minj - 1 : ncrd - minj + maxj] = crd_numpy[: maxj + 1]
-            op1_right_of_op2 = on_right_side(op1, op2)
-            if (right_bank and op1_right_of_op2) or (
-                (not right_bank) and (not op1_right_of_op2)
-            ):
-                op = op2
-            else:
-                op = op1
-            banklines_list[i] = LineString(op)
-        else:
-            if minj < maxj:
-                banklines_list[i] = LineString(bl.coords[minj : maxj + 1])
-            else:  # minj > maxj
-                banklines_list[i] = LineString(bl.coords[maxj : minj + 1][::-1])
-        lengths[i] = maxloc - minloc
-
-    # print("select lines by length")
-    while True:
-        maxl = lengths.max()
-        if maxl == 0:
-            break
-        iarray = numpy.nonzero(lengths == maxl)
-        i = iarray[0][0]
-        # print("i={}, length={}, minlocs={}, maxlocs={}".format(i, lengths[i], minlocs[i], maxlocs[i]))
-        keep[i] = True
-        # remove lines that are a subset
-        lengths[(minlocs >= minlocs[i]) & (maxlocs <= maxlocs[i])] = 0
-        # print("lengths[i] set to {}".format(lengths[i]))
-        # if line partially overlaps ... but stick out on the high side
-        jarray = numpy.nonzero(
-            (minlocs > minlocs[i]) & (minlocs < maxlocs[i]) & (maxlocs > maxlocs[i])
-        )[0]
-        if jarray.size > 0:
-            for j in jarray:
-                bl = banklines_list[j]
-                kmax = len(bl.coords) - 1
-                for k, p in enumerate(bl.coords):
-                    if k == kmax:
-                        # a line string of a single point would remain
-                        lengths[j] = 0
-                        break
-                    loc = xykm.project(Point(p))
-                    if loc >= maxlocs[i]:
-                        banklines_list[j] = LineString(bl.coords[k:])
-                        minlocs[j] = loc
-                        break
-        # if line partially overlaps ... but stick out on the low side
-        jarray = numpy.nonzero(
-            (minlocs < minlocs[i]) & (maxlocs > minlocs[i]) & (maxlocs < maxlocs[i])
-        )[0]
-        if jarray.size > 0:
-            for j in jarray:
-                bl = banklines_list[j]
-                kmax = len(bl.coords) - 1
-                for k, p in zip(range(-1, -kmax, -1), bl.coords[:-1][::-1]):
-                    if k == kmax + 1:
-                        # a line string of a single point would remain
-                        lengths[j] = 0
-                        break
-                    loc = xykm.project(Point(p))
-                    if loc <= minlocs[i]:
-                        banklines_list[j] = LineString(bl.coords[:k])
-                        maxlocs[j] = loc
-                        break
-
-    # select banks in order of projected length
-    idx = numpy.argsort(minlocs[keep])
-    idx2 = numpy.nonzero(keep)[0]
-    new_bank_coords = []
-    for i in idx2[idx]:
-        new_bank_coords.extend(banklines_list[i].coords)
-    bank = LineString(new_bank_coords)
-
-    return bank
-
-
-def poly_to_line(
-    nnodes: int,
-    x: numpy.ndarray,
-    y: numpy.ndarray,
-    wet_node: numpy.ndarray,
-    h_node: numpy.ndarray,
-    h0: float,
-):
-    """
-    Detect the bank line segments inside an individual face of arbitrary (convex) polygonal shape.
-
-    Arguments
-    ---------
-    nnode : int
-        Number of nodes of mesh face.
-    x : numpy.ndarray
-        Array of x-coordinates of the nodes making up the mesh face.
-    y : numpy.ndarray
-        Array of y-coordinates of the nodes making up the mesh face.
-    wet_node : numpy.ndarray
-        Array of booleans indicating whether nodes are wet.
-    h_node : numpy.ndarray
-        Array of water depths (negative for dry) at the mesh nodes.
-    h0 : float
-        Critical water depth for determining the banks.
-    
-    Results
-    -------
-    lines : Optional[...]
-        Optional bank line segments detected within the mesh face.
-    """
-    Lines = [None] * (nnodes - 2)
-    for i in range(nnodes - 2):
-        iv = [0, i + 1, i + 2]
-        nwet = sum(wet_node[iv])
-        if nwet == 1 or nwet == 2:
-            # print("x: ",x[iv]," y: ",y[iv], " w: ", wet_node[iv], " d: ", h_node[iv])
-            Lines[i] = tri_to_line(x[iv], y[iv], wet_node[iv], h_node[iv], h0)
-    Lines = [line for line in Lines if not line is None]
-    if len(Lines) == 0:
-        return None
-    else:
-        multi_line = MultiLineString(Lines)
-        merged_line = line_merge(multi_line)
-        return merged_line
-
-
-def tri_to_line(
-    x: numpy.ndarray,
-    y: numpy.ndarray,
-    wet_node: numpy.ndarray,
-    h_node: numpy.ndarray,
-    h0: float,
-):
-    """
-    Detect the bank line segments inside an individual triangle.
-
-    Arguments
-    ---------
-    x : numpy.ndarray
-        Array of x-coordinates of the nodes making up the mesh face.
-    y : numpy.ndarray
-        Array of y-coordinates of the nodes making up the mesh face.
-    wet_node : numpy.ndarray
-        Array of booleans indicating whether nodes are wet.
-    h_node : numpy.ndarray
-        Array of water depths (negative for dry) at the mesh nodes.
-    h0 : float
-        Critical water depth for determining the banks.
-        
-    Returns
-    -------
-    Line : Optional[]
-        Optional bank line segment detected within the triangle.
-    """
-    if wet_node[0] and wet_node[1]:
-        A = 0
-        B = 2
-        C = 1
-        D = 2
-    elif wet_node[0] and wet_node[2]:
-        A = 0
-        B = 1
-        C = 2
-        D = 1
-    elif wet_node[0]:
-        A = 0
-        B = 1
-        C = 0
-        D = 2
-    elif wet_node[1] and wet_node[2]:
-        A = 2
-        B = 0
-        C = 1
-        D = 0
-    elif wet_node[1]:
-        A = 1
-        B = 0
-        C = 1
-        D = 2
-    else:  # wet_node[2]
-        A = 2
-        B = 0
-        C = 2
-        D = 1
-    facAB = (h_node[A] - h0) / (h_node[A] - h_node[B])  # large facAB -> close to B
-    xl = x[A] + facAB * (x[B] - x[A])
-    yl = y[A] + facAB * (y[B] - y[A])
-    facCD = (h_node[C] - h0) / (h_node[C] - h_node[D])  # large facCD -> close to D
-    xr = x[C] + facCD * (x[D] - x[C])
-    yr = y[C] + facCD * (y[D] - y[C])
-    if xl == xr and yl == yr:
-        Line = None
-    else:
-        Line = LineString([[xl, yl], [xr, yr]])
-    return Line
 
 
 def enlarge(
